requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=6.0
pandas>=2.0.0
Pillow>=10.0.0
scipy>=1.11.0
lapx>=0.5.5
pygame>=2.5.0
requests>=2.31.0
tqdm>=4.66.0
colorama>=0.4.6
